Index Overloading, Position Change, and Lines
Created overloads for the FindIndex function to be able to work with
different types of lists.
Adjusted how the tranform.position of objects is calculated. Moved it
from being done in a FOR loop to a FOREACH statement.
Wrote a function for rendering lines to represent the spring-dampers.
Commented out all implementation for the function due to cost.
Particles now spawn as a child of an empty game object.